moved detailsviewhelper to the biscuitsviewhelper class
