+++ biscuitdetails.php
@@ -3,12 +3,16 @@ require_once './vendor/autoload.php';
 
 use BisquidsTin\Hydrators\BiscuitsHydrator;
 use BisquidsTin\Utilities\DB;
-use BisquidsTin\ViewHelpers\DetailsViewHelper;
+use BisquidsTin\ViewHelpers\BiscuitsViewHelper;
 
-$id = $_POST['id'];
-$db = DB::getDB();
-$biscuit = BiscuitsHydrator::getBiscuitById($db, $id);
-$biscuitDetailsDisplay = DetailsViewHelper::displayBiscuitDetails($biscuit);
+if (isset($_GET['id']) && $_GET['id'] !== '') {
+    $id = $_GET['id'];
+    $db = DB::getDB();
+    $biscuit = BiscuitsHydrator::getBiscuitById($db, $id);
+    $biscuitDetailsDisplay = BiscuitsViewHelper::displayBiscuitDetails($biscuit);
+} else {
+    header('Location: index.php');
+}
 
 ?>
 <html lang="en-gb">
